requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.0
python-wordpress-xmlrpc>=2.2
cliff>=1.14.0
Markdown>=2.2.1
